% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{createVennObj}
\alias{createVennObj}
\title{Create nVennObj from scratch}
\usage{
createVennObj(nSets = 1, sNames = NULL, sSizes = NULL)
}
\arguments{
\item{nSets}{Number of sets.}

\item{sNames}{List of names.}

\item{sSizes}{List of sizes for all the regions (from `0` to `2**nSets - 1`).}
}
\value{
nVennObj with set information. To plot, it must be sent to `toVenn`. Sending it to `showSVG`
will render the diagram before simulation.
}
\description{
Create nVennObj from scratch
}
