% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getVennSetNames}
\alias{getVennSetNames}
\title{Gets the names of the sets}
\usage{
getVennSetNames(nVennObj)
}
\arguments{
\item{nVennObj}{nVennR object generated with \code{\link[=nVennDiagram]{nVennDiagram()}}.}
}
\value{
List of set names.
}
\description{
Gets the names of the sets
}
\examples{
myv <- nVennDiagram(list(Set1=c("a", "b", "c"), Set2=c("a", "c", "d")), verbose=F)
getVennSetNames(myv)
}
