% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nVenn2.R
\name{plotSVG}
\alias{plotSVG}
\title{Plot nVenn diagram}
\usage{
plotSVG(nVennObj, outFile = "", systemShow = F)
}
\arguments{
\item{nVennObj}{nVenn object.}

\item{outFile}{Path to export the SVG figure. If empty, the figure is not
exported.}

\item{systemShow}{If true, and the system has a default SVG-editing program,
opens the figure in the default editor}
}
\description{
The nVenn diagram is generated in the plot window by default. If this is not
possible, parameters outFile and systemShow may be used to export the SVG 
figure.
}
