% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotateVenn}
\alias{rotateVenn}
\title{Rotate a diagram}
\usage{
rotateVenn(nVennObj, angle)
}
\arguments{
\item{nVennObj}{nVennR object generated with \code{\link[=nVennDiagram]{nVennDiagram()}}.}

\item{angle}{Angle of rotation in degrees. Positive values rotate
counterclockwise.}
}
\value{
String with svg code.
}
\description{
Rotate a diagram
}
\details{
The code returned by the funciton can be saved to a file and then
edited with vectorial image
software. This can be done directly with \code{\link[=nVennDiagram]{nVennDiagram()}} or
\code{\link[=plotVenn]{plotVenn()}} by using the \code{outFile} param.
}
\examples{
myv <- nVennDiagram(list(Set1=c("a", "b", "c"), Set2=c("a", "c", "d")), verbose=F)
getVennSvg(myv)
}
