% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nVenn2.R
\name{setVennColor}
\alias{setVennColor}
\title{Change a set color}
\usage{
setVennColor(nVennObj, setName, color)
}
\arguments{
\item{nVennObj}{nVenn object created with \code{\link[=nVennDiagram]{nVennDiagram()}}.}

\item{setName}{Name of the set.}

\item{color}{New color. It is important to notice that this parameter must be
a valid color in SVG format. The value is not checked, and therefore an
incorrect value will break the plot in \code{\link[=plotVenn]{plotVenn()}}.}
}
\value{
nVenn object with the change in color for the set. This function
does not plot the result, \code{\link[=plotVenn]{plotVenn()}} needs to be called afterwards.
}
\description{
Change a set color
}
\examples{
myv <- nVennDiagram(list(Set1=c("a", "b", "c"), Set2=c("a", "c", "d")), verbose=F)
myv <- setVennColor(myv, "Set2", "black")
plotVenn(myv)
myv <- setVennColor(myv, "Set1", "#ffff00")
plotVenn(myv)
}
