% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nVenn2.R
\name{setVennOpts}
\alias{setVennOpts}
\title{Set graphical parameters of an nVenn object diagram.}
\usage{
setVennOpts(
  nVennObj,
  opacity = 0.4,
  fontSize = 12,
  lineWidth = 1,
  palette = 0,
  showRegions = T,
  showWeights = T
)
}
\arguments{
\item{nVennObj}{nVenn object generated with \code{\link[=nVennDiagram]{nVennDiagram()}}.}

\item{opacity}{Opacity of sets, between 0 (completely transparent)
and 1 (completely opaque). Defaults to 0.4.}

\item{fontSize}{Size of the font for the numbers of elements (weights) in each
region. It also indirectly controls the font sizes of the
region descriptions, if shown (showRegions=T). Defaults to
10.}

\item{lineWidth}{Width of the lines defining each set. The value will be
rounded to the closest integer. Defaults to 1.}

\item{palette}{Color palette to use to fill the sets. Integer from 0 to 3.
Defaults to 0.}

\item{showRegions}{If true (default), show region descriptions.}

\item{showWeights}{If true (default), show number of elements in each region.}
}
\value{
nVennObj with graphical options set.
}
\description{
The function returns an object whose parameters will change the way the
diagram is plotted. To see the changes, it is necessary to run \code{\link[=plotVenn]{plotVenn()}}
afterwards.
}
\examples{
myv <- nVennDiagram(exampledf)
myv <- setVennOpts(myv, showRegions=F, opacity=0.2, lineWidth=2)
plotVenn(myv)
}
\seealso{
\link{plotVenn} for plotting the resulting diagram.
}
