% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{setVennRegion}
\alias{setVennRegion}
\title{Get elements in a region}
\usage{
setVennRegion(nVennObj, region, value)
}
\arguments{
\item{nVennObj}{Object describing an nVenn job.}

\item{region}{Description of the region. This can be a vector with the names of the groups the region
belongs to or a vector describing whether the region belongs to each set in order (i. e., c(1, 0, 0)
means the region belongs to set 1 and does not belong to sets 2 and 3).}

\item{value}{Size of the region.}
}
\value{
Modified nVennObj
}
\description{
Get elements in a region
}
