% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{showSVG}
\alias{showSVG}
\title{Show Venn diagram. Automatically called from toVenn.}
\usage{
showSVG(mySVG, opacity = 0.4, outFile = "", systemShow = FALSE)
}
\arguments{
\item{mySVG}{SVG code defining the diagram. Can be retrieved from toVenn.}

\item{opacity}{Fill opacity for the sets. Defaults to 0.4.}

\item{outFile}{File name to save SVG figure. If empty, a temp file will be created and
shown in the viewer, if possible.}

\item{systemShow}{Show the result in the system SVG viewer (i. e., Inkscape).}
}
\value{
Nothing. Creates a Venn diagram in svg as a side effect.
}
\description{
Show Venn diagram. Automatically called from toVenn.
}
