% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{toVenn}
\alias{toVenn}
\title{This function is deprecated, only kept for backward compatibility.}
\usage{
toVenn(..., draw = TRUE)
}
\arguments{
\item{...}{One list or vector (possibly mixed) per set. If the input
is a list with a name, that name will be used for the legend.}

\item{draw}{Show Venn diagram in the viewer as a side effect. Defaults to true.}
}
\value{
SVG code for the Venn diagram.
}
\description{
Please, use plotVenn instead. Please, notice that input and output are different.
}
\examples{
set1 <- c('a', 'b', 'c')
set2 <- c('e', 'f', 'c')
set3 <- c('c', 'b', 'e')
mySVG <- toVenn(set1, set2, set3)
oldShowSVG(mySVG=mySVG, opacity=0.2)
}
