% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nVenn2.R
\docType{data}
\name{exampledf}
\alias{exampledf}
\title{Example data frame.}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{Employee}{Employee ID}
  \item{SAS}{Employee uses SAS}
  \item{Python}{Employee uses Python}
  \item{R}{Employee uses R}
}
}
\source{
\url{https://stackoverflow.com/questions/49471565/transforming-data-to-create-generalized-quasi-proportional-venn-diagrams-using}
}
\usage{
exampledf
}
\description{
A dataset containing programming preferences from 18 employees. This data set was
provided by user Krantz to inquire about nVennR
}
\keyword{datasets}
