% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getVennRegion}
\alias{getVennRegion}
\title{Get elements in a region}
\usage{
getVennRegion(nVennObj, region)
}
\arguments{
\item{nVennObj}{Object describing an nVenn job.}

\item{region}{Description of the region. This can be a vector with the names of the groups the region
belongs to or a vector describing whether the region belongs to each set in order (i. e., c(1, 0, 0)
means the region belongs to set 1 and does not belong to sets 2 and 3).}
}
\value{
list of the elements belonging to the specified region
}
\description{
Get elements in a region
}
