% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getVennRegion}
\alias{getVennRegion}
\title{Get elements in a region}
\usage{
getVennRegion(nVennObj, region)
}
\arguments{
\item{nVennObj}{Object describing an nVenn job.}

\item{region}{Description of the region.}
}
\value{
list of the elements belonging to the specified region
}
\description{
Get elements in a region
}
