% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{listVennRegions}
\alias{listVennRegions}
\title{List elements in every region}
\usage{
listVennRegions(nVennObj, na.rm = T)
}
\arguments{
\item{nVennObj}{Object to list.}

\item{na.rm}{If true, empty regions are not listed.}
}
\description{
List elements in every region
}
