% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nVennDiagram}
\alias{nVennDiagram}
\title{Creates nVenn plot}
\usage{
nVennDiagram(desc, verbose = TRUE, byCol = 0L)
}
\arguments{
\item{desc}{Description of sets, either as a list of lists or as text.}

\item{v2}{verbose If true, shows messages as the nVenn plot is created.}
}
\value{
nVenn object. As a side effect, shows the nVenn plot.
}
\description{
Creates nVenn plot
}
