% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nVenn2.R
\name{plotSVG}
\alias{plotSVG}
\title{Title Plot the current Venn diagram.}
\usage{
plotSVG(systemShow = F)
}
\arguments{
\item{systemShow}{If true, it also displays the result in the system-defined editor.}
}
\description{
Title Plot the current Venn diagram.
}
