% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plotVenn}
\alias{plotVenn}
\title{Create Venn diagram using the nVenn algorithm.}
\usage{
plotVenn(..., sNames = c())
}
\arguments{
\item{...}{One list or vector (possibly mixed) per set. If the input
is a list with a name, that name will be used for the legend. If not, names can be
provided with \code{sNames}.}

\item{sNames}{List of set names, in the same order as the input lists.}
}
\value{
SVG code for the Venn diagram. As a side effect, The result is drawn in the
plot window.
}
\description{
This algorithm is based on a simulation
that compacts the figure. To avoid clogging the system, the simulation stops every 7000
cycles and asks the user if it sould go on. Only answering 'y' continues. Once the
diagram is compact, it will be slightly embellished. This function is different in use and
output from toVenn, which will probably deprecated.plotVenn
}
\examples{
set1 <- c('a', 'b', 'c')
set2 <- c('e', 'f', 'c')
set3 <- c('c', 'b', 'e')
mySVG <- plotVenn(set1, set2, set3, sNames=c("One", "Two", "Three"))
showSVG(mySVG=mySVG, opacity=0.2)
}
