% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotateVenn}
\alias{rotateVenn}
\title{Rotate a diagram}
\usage{
rotateVenn(nVennObj, angle, plot = TRUE)
}
\arguments{
\item{nVennObj}{nVennR object generated with \code{\link[=nVennDiagram]{nVennDiagram()}}.}

\item{angle}{Angle of rotation in degrees. Positive values rotate
counterclockwise.}

\item{plot}{If true (default), prints the diagram after the rotation.}
}
\value{
nVenn object.
}
\description{
Rotate a diagram
}
