% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotateVenn}
\alias{rotateVenn}
\title{Rotate a diagram}
\usage{
rotateVenn(nVennObj, angle)
}
\arguments{
\item{nVennObj}{nVennR object generated with \code{\link[=nVennDiagram]{nVennDiagram()}}.}

\item{angle}{Angle of rotation in degrees. Positive values rotate
counterclockwise.}
}
\value{
nVenn object. The function does not plot the diagram. It is
necessary to run \code{\link[=plotVenn]{plotVenn()}} for that.
}
\description{
Rotate a diagram
}
