% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{toVenn}
\alias{toVenn}
\title{Create Venn diagram using the nVenn algorithm.}
\usage{
toVenn(...)
}
\arguments{
\item{...}{One list or vector (possibly mixed) per set. If the input
is a list with a name, that name will be used for the legend}
}
\value{
Nothing. Creates a Venn diagram in svg as a side effect.
}
\description{
Create Venn diagram using the nVenn algorithm.
}
\examples{
set1 <- c('a', 'b', 'c')
set2 <- c('e', 'f', 'c')
set3 <- c('c', 'b', 'e')
toVenn(set1, set2, set3)
}
